single project form component with edit
