+++ packages/renderer/src/views/view-summary.vue
@@ -6,7 +6,7 @@ import BaseTemplateCard from "../components/base-template-card.vue";
 import BaseCardContent from "../components/base-card-content.vue";
 import BaseButtonClick from "../components/base-button-click.vue";
 import BaseModal from "../components/base-modal.vue";
-import FormProjectCreateModal from "../components/form-project-create-modal.vue";
+import FormProject from "../components/form-project.vue";
 import FormProjectDeleteModal from "../components/form-project-delete-modal.vue";
 import NotificationDot from "../components/notification-dot.vue";
 import ButtonEllipsis from "../components/button-ellipsis.vue";
@@ -17,11 +17,11 @@ import FormGoal from "../components/form-goal.vue";
 import FormGoalManageModal from "../components/form-goal-manage-modal.vue";
 
 // TODO:
-// If completed or archived, no longer able to add/edit content
+// If completed or archived, no longer able to add/edit content or Project details
 
 const store = inject("store") as Store;
 
-const isEditFormModalActive = ref<boolean>(false);
+const isEditProjectModalActive = ref<boolean>(false);
 const isCreateGoalFormModalActive = ref<boolean>(false);
 const isManageGoalModalActive = ref<boolean>(false);
 const isDeleteModalActive = ref<boolean>(false);
@@ -39,14 +39,6 @@ function openNotesColumn(): void {
       column.isActive = !column.isActive;
     }
   });
-}
-
-function openEditProjectModal(): void {
-  console.log("OPEN EDIT PROJECT MODAL");
-}
-
-function openEditGoalModal(): void {
-  console.log("OPEN EDIT GOAL MODAL");
 }
 
 function toggleProjectComplete(): void {
@@ -124,7 +116,7 @@ const ellipsisMenuArchivedText = computed(() => {
     </template>
 
     <template #ellipsis-menu>
-      <button-ellipsis-item @click="openEditProjectModal">
+      <button-ellipsis-item @click="isEditProjectModalActive = true">
         Edit Project
       </button-ellipsis-item>
 
@@ -213,10 +205,13 @@ const ellipsisMenuArchivedText = computed(() => {
     @handle-delete-modal-close="isDeleteModalActive = false"
   />
 
-  <form-project-create-modal
-    :is-form-modal-active="isEditFormModalActive"
-    @close-modal="isEditFormModalActive = false"
-  />
+  <base-modal
+    :is-modal-active="isEditProjectModalActive"
+    @close-modal="isEditProjectModalActive = false"
+  >
+    <template #header> Edit Project </template>
+    <form-project :current-project="{ ...activeProject }" />
+  </base-modal>
 
   <base-modal
     :is-modal-active="isCreateGoalFormModalActive"
